+++ tests/Framework/Tests/IntegrationTest.php
@@ -14,11 +14,11 @@ declare(strict_types=1);
 namespace Rekalogika\DomainEvent\Tests\Framework\Tests;
 
 use Psr\EventDispatcher\EventDispatcherInterface;
-use Rekalogika\DomainEvent\Constants;
+use Rekalogika\DomainEvent\DependencyInjection\Constants;
 use Rekalogika\DomainEvent\Doctrine\DoctrineEventListener;
 use Rekalogika\DomainEvent\Doctrine\DomainEventAwareManagerRegistry;
 use Rekalogika\DomainEvent\DomainEventAwareEntityManagerInterface;
-use Rekalogika\DomainEvent\DomainEventReaper;
+use Rekalogika\DomainEvent\Doctrine\DomainEventReaper;
 use Rekalogika\DomainEvent\ImmediateDomainEventDispatcherInstaller;
 
 final class IntegrationTest extends DomainEventTestCase